+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Karafka Framework Changelog
 
 ## 2.4.11 (Unreleased)
+- [Enhancement] Validate envelope target topic type for Scheduled Messages.
 - [Enhancement] Support for enqueue_after_transaction_commit in rails active job.
 - [Fix] Fix invalid reference to AppSignal version.